inherit common methods from entity::base
