+++ lib/evertils/common/entity/stack.rb
@@ -1,15 +1,7 @@
-require 'evertils/common/authentication'
-
 module Evertils
   module Common
     module Entity
-      class Stack
-        
-        def initialize
-          @evernote = Authentication.new.store
-
-          self
-        end
+      class Stack < Entity::Base
 
         def create_from_yml(full_path)
           begin
@@ -33,14 +25,6 @@ module Evertils
             end
           rescue ArgumentError => e
             puts e.message
-          end
-        end
-
-        private
-
-        def has_required_fields(hash, required)
-          hash.keys.each do |key|
-            required.include? key
           end
         end
