+++ lib/evertils/common/entity/tags.rb
@@ -1,15 +1,7 @@
-require 'evertils/common/authentication'
-
 module Evertils
   module Common
     module Entity
-      class Tags
-        
-        def initialize
-          @evernote = Authentication.new.store
-
-          self
-        end
+      class Tags < Entity::Base
 
         def all
           @evernote.listTags(Evertils::Common::EVERNOTE_DEVELOPER_TOKEN)
